fix(env): harden filtered subprocess state
